Bart base training - first pass
It pulls the json files - asset train and test - from my driver since i was working in collab. Please change the path according to where your file is located! And we can make it extensible later for sure. - and ive added a testing block at the end but feel free to remove it and test on the entire file